requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
tensorboard>=2.8.0
PyYAML>=6.0
pandas>=1.3.0
jupyter>=1.0.0
notebook>=6.4.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0 